Remove logic from pay leave for parents flow for old due dates
The flow did handle due dates prior to 2015-04-05, I guess because the
policy changed at this time. I believe this smart answer only needs to
work going forward, so this complexity can be removed.

As well as changing the flow, I've removed all outcomes that don't
look to be reachable, and all tests that mention removed outcomes.

TODO: Some tests still fail, and this needs to be addressed somehow...